feat: Add page mode filter (working in progress)
