--- thermal_parser/src/utils/mod.rs
@@ -1,1 +1,0 @@
-pub mod page_mode_filter;
