--- thermal_parser/src/utils/page_mode_filter/mod.rs
@@ -1,1 +1,0 @@
-pub mod filter;
